requirements.txt
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
matplotlib>=3.7.0
Pillow>=10.0.0
pandas>=2.0.0
seaborn>=0.12.0
tqdm>=4.65.0
jupyter>=1.0.0
notebook>=7.0.0
torch>=2.0.0
torchvision>=0.15.0